// app/(manager)/(tabs)/types.ts
export interface UserDataProps {
  id: string;
  name: string;
  email: string;
  phone: string;
  homeAddress: string;
  dateOfBirth: string;
  Language: string;
  identityVerificationStatus: string;
  backgroundCheckStatus: string;
  approvalStatus: string;
  description: string;
  rating: string,
  assurance: string,
  urlToImage: any,
  imageStatus: string;
  IDCardStatus: string;
  PassportStatus: string;
  otherCertificateStatus: string;
}